+++ src/services/Elements.php
@@ -1862,10 +1862,12 @@ class Elements extends Component
             if (is_array($path)) {
                 $criteria = $path['criteria'] ?? $path[1] ?? null;
                 $count = $path['count'] ?? ArrayHelper::remove($criteria, 'count', false);
+                $when = $path['when'] ?? null;
                 $path = $path['path'] ?? $path[0];
             } else {
                 $criteria = null;
                 $count = false;
+                $when = null;
             }
 
             // Split the path into the top segment and subpath
@@ -1905,6 +1907,10 @@ class Elements extends Component
                 } else {
                     $plan->all = true;
                 }
+
+                if ($when !== null) {
+                    $plan->when = $when;
+                }
             } else {
                 // We are for sure going to need to query the elements
                 $plan->all = true;
@@ -1914,6 +1920,7 @@ class Elements extends Component
                     'path' => $subpath,
                     'criteria' => $criteria,
                     'count' => $count,
+                    'when' => $when,
                 ];
             }
         }
@@ -1963,9 +1970,16 @@ class Elements extends Component
             $this->trigger(self::EVENT_BEFORE_EAGER_LOAD_ELEMENTS, $event);
 
             foreach ($event->with as $plan) {
+                // Filter out any elements that the plan doesn't like
+                if ($plan->when !== null) {
+                    $filteredElements = array_values(array_filter($elements, $plan->when));
+                } else {
+                    $filteredElements = $elements;
+                }
+
                 // Get the eager-loading map from the source element type
                 /** @var ElementInterface|string $elementType */
-                $map = $elementType::eagerLoadingMap($elements, $plan->handle);
+                $map = $elementType::eagerLoadingMap($filteredElements, $plan->handle);
 
                 if ($map === null) {
                     // Null means to skip eager-loading this segment
@@ -2036,7 +2050,7 @@ class Elements extends Component
                     }
 
                     // Loop through the source elements and count up their targets
-                    foreach ($elements as $sourceElement) {
+                    foreach ($filteredElements as $sourceElement) {
                         $count = 0;
                         if (!empty($targetElementIdCounts) && isset($targetElementIdsBySourceIds[$sourceElement->id])) {
                             foreach (array_keys($targetElementIdsBySourceIds[$sourceElement->id]) as $targetElementId) {
@@ -2055,7 +2069,7 @@ class Elements extends Component
                 $targetElements = [];
 
                 // Tell the source elements about their eager-loaded elements
-                foreach ($elements as $sourceElement) {
+                foreach ($filteredElements as $sourceElement) {
                     $targetElementIdsForSource = [];
                     $targetElementsForSource = [];